requirements.txt
discord.py==0.16.8
aiohttp
websockets
chardet
google-api-python-client
requests
